refactor(apps/api): improve lists dir
