+++ apps/api/src/plugins/invites/get/getUserInvites.test.ts
@@ -1,7 +1,8 @@
 import { prisma } from "@hominem/db";
 import { FastifyInstance } from "fastify";
-import { createServer } from "../../../server";
-import { mockAuthSession } from "../../../../test/utils";
+
+import { createServer } from "@app/server";
+import { mockAuthSession } from "@test/utils";
 
 describe("GET /invites", () => {
   let server: FastifyInstance;
