+++ apps/api/src/plugins/lists/lists.test.ts
@@ -1,8 +1,8 @@
 import { prisma } from "@hominem/db";
 import { FastifyInstance } from "fastify";
 
-import { mockAuthSession } from "../../../test/utils";
-import { createServer } from "../../server";
+import { createServer } from "@app/server";
+import { mockAuthSession } from "@test/utils";
 
 describe("/lists", () => {
   let server: FastifyInstance;
